+++ imod/tests/test_msw/test_sprinkling.py
@@ -62,15 +62,14 @@ def test_simple_model(fixed_format_parser):
     cellids = derive_cellid_from_points(svat, well_x, well_y, well_layer)
     well = Mf6Wel(cellids, well_rate)
 
-    coupler_mapping = msw.Sprinkling(
+    sprinkling = msw.Sprinkling(
         max_abstraction_groundwater,
         max_abstraction_surfacewater,
-        well,
     )
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        coupler_mapping.write(output_dir, index, svat)
+        sprinkling.write(output_dir, index, svat, None, well)
 
         results = fixed_format_parser(
             output_dir / msw.Sprinkling._file_name,
@@ -79,11 +78,11 @@ def test_simple_model(fixed_format_parser):
 
     assert_equal(results["svat"], np.array([1, 2, 3, 4]))
     assert_almost_equal(
-        results["max_abstraction_groundwater_m3_d"],
+        results["max_abstraction_groundwater"],
         np.array([100.0, 300.0, 100.0, 200.0]),
     )
     assert_almost_equal(
-        results["max_abstraction_surfacewater_m3_d"],
+        results["max_abstraction_surfacewater"],
         np.array([100.0, 300.0, 100.0, 200.0]),
     )
     assert_equal(results["layer"], np.array([3, 1, 3, 2]))
@@ -136,15 +135,14 @@ def test_simple_model_1_subunit(fixed_format_parser):
     cellids = derive_cellid_from_points(svat, well_x, well_y, well_layer)
     well = Mf6Wel(cellids, well_rate)
 
-    coupler_mapping = msw.Sprinkling(
+    sprinkling = msw.Sprinkling(
         max_abstraction_groundwater,
         max_abstraction_surfacewater,
-        well,
     )
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        coupler_mapping.write(output_dir, index, svat)
+        sprinkling.write(output_dir, index, svat, None, well)
 
         results = fixed_format_parser(
             output_dir / msw.Sprinkling._file_name,
@@ -153,11 +151,11 @@ def test_simple_model_1_subunit(fixed_format_parser):
 
     assert_equal(results["svat"], np.array([1, 2]))
     assert_almost_equal(
-        results["max_abstraction_groundwater_m3_d"],
+        results["max_abstraction_groundwater"],
         np.array([100.0, 300.0]),
     )
     assert_almost_equal(
-        results["max_abstraction_surfacewater_m3_d"],
+        results["max_abstraction_surfacewater"],
         np.array([100.0, 300.0]),
     )
     assert_equal(results["layer"], np.array([3, 2]))